requirements.txt
ultralytics>=8.0.0
easyocr>=1.6.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
PyYAML>=6.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0